Add sequence rename in inspector
